+++ src/expression/logical/index.ts
@@ -34,8 +34,8 @@ export abstract class Logical implements Evaluable {
    */
   abstract simplify(
     ctx: Context,
-    strictKeys?: false | string[],
-    optionalKeys?: false | string[]
+    strictKeys?: string[],
+    optionalKeys?: string[]
   ): Result | Evaluable
 
   /**