+++ src/index.ts
@@ -98,10 +98,10 @@ class Engine {
    *
    * @param {ExpressionInput} exp  Raw expression.
    * @param {Context} context Evaluation data context.
-   * @param {false | string[]} strictKeys keys to be considered present even if they are not present in the context
+   * @param {string[]} strictKeys keys to be considered present even if they are not present in the context
    *  Defaults to false
-   * @param {false | string[]} optionalKeys keys to be considered not present unless they are also in `strictKeys`;
-   *  when `strictKeys` is set to `false` and `optionalKeys` is an array, every key that is not in `optionalKeys`
+   * @param {string[]} optionalKeys keys to be considered not present unless they are also in `strictKeys`;
+   *  when `strictKeys` is `undefined` and `optionalKeys` is an array, every key that is not in `optionalKeys`
    *  is considered to be present and thus will be evaluated
    *  Defaults to false
    * @returns {Inpunt | boolean}
@@ -109,8 +109,8 @@ class Engine {
   simplify(
     exp: ExpressionInput,
     context: Context,
-    strictKeys: false | string[] = false,
-    optionalKeys: false | string[] = false
+    strictKeys?: string[],
+    optionalKeys?: string[]
   ): Input | boolean {
     const result = this.parse(exp).simplify(context, strictKeys, optionalKeys)
     if (isEvaluable(result)) {