+++ types/operand/index.d.ts
@@ -13,7 +13,7 @@ export declare abstract class Operand implements Evaluable {
     /**
      * {@link Evaluable.simplify}
      */
-    abstract simplify(ctx: Context, strictKeys?: false | string[], optionalKeys?: false | string[]): Result | Evaluable;
+    abstract simplify(ctx: Context, strictKeys?: string[], optionalKeys?: string[]): Result | Evaluable;
     /**
      * {@link Evaluable.serialize}
      */